flightTask/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164373/work
certifi==2022.6.15
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1656782830073/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1655906222726/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
conda==4.13.0
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1649385069444/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1657173995269/work
distlib==0.3.5
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
filelock==3.8.0
frozenlist==1.3.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
multidict==6.0.2
mysqlclient==2.1.1
pipenv==2022.8.24
platformdirs==2.5.2
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1649384802055/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
PyMySQL==1.0.2
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857263093/work
python-weather==0.4.1
pytz==2022.2.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1656534056640/work
ruamel-yaml-conda @ file:///home/conda/feedstock_root/build_artifacts/ruamel_yaml_1653464393916/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sqlparse==0.4.2
tk==0.1.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1649051611147/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
virtualenv==20.16.3
virtualenv-clone==0.5.7
yarl==1.8.1
